+++ backend/libraries/timer_job_queues/src/grouped_timer_job_queue.rs
@@ -15,21 +15,34 @@ pub struct GroupedTimerJobQueue<T: TimerJobItemGroup> {
 }
 
 impl<T: TimerJobItemGroup> GroupedTimerJobQueue<T> {
-    pub fn new(max_concurrency: usize) -> Self {
+    pub fn new(max_concurrency: usize, defer_processing: bool) -> Self {
         Self {
             inner: Rc::new(Mutex::new(GroupedTimerJobQueueInner {
                 queue: VecDeque::new(),
                 items_map: BTreeMap::new(),
                 in_progress: BTreeSet::new(),
                 max_concurrency,
+                defer_processing,
                 timer_id: None,
             })),
             phantom: PhantomData,
         }
     }
 
+    pub fn max_concurrency(&self) -> usize {
+        self.within_lock(|i| i.max_concurrency)
+    }
+
     pub fn set_max_concurrency(&self, value: usize) {
         self.within_lock(|i| i.max_concurrency = value)
+    }
+
+    pub fn defer_processing(&self) -> bool {
+        self.within_lock(|i| i.defer_processing)
+    }
+
+    pub fn set_defer_processing(&self, value: bool) {
+        self.within_lock(|i| i.defer_processing = value)
     }
 
     pub fn clear(&self) {
@@ -60,6 +73,7 @@ struct GroupedTimerJobQueueInner<K: Clone + Ord, I> {
     items_map: BTreeMap<K, VecDeque<I>>,
     in_progress: BTreeSet<K>,
     max_concurrency: usize,
+    defer_processing: bool,
     #[serde(skip)]
     timer_id: Option<TimerId>,
 }
@@ -68,21 +82,29 @@ impl<T: TimerJobItemGroup + 'static> GroupedTimerJobQueue<T>
 where
     <T as TimerJobItemGroup>::Key: Clone + Ord,
 {
-    pub fn enqueue(&self, grouping_key: T::Key, item: T::Item) {
-        self.enqueue_many(grouping_key, vec![item])
+    pub fn push(&self, grouping_key: T::Key, item: T::Item) {
+        self.push_many(grouping_key, vec![item])
     }
 
-    pub fn enqueue_many(&self, grouping_key: T::Key, items: Vec<T::Item>) {
-        self.within_lock(|i| match i.items_map.entry(grouping_key.clone()) {
-            Vacant(e) => {
-                e.insert(VecDeque::from(items));
-                i.queue.push_back(grouping_key);
+    pub fn push_many(&self, grouping_key: T::Key, items: Vec<T::Item>) {
+        let defer_processing = self.within_lock(|i| {
+            match i.items_map.entry(grouping_key.clone()) {
+                Vacant(e) => {
+                    e.insert(VecDeque::from(items));
+                    i.queue.push_back(grouping_key);
+                }
+                Occupied(mut e) => {
+                    e.get_mut().extend(items);
+                }
             }
-            Occupied(mut e) => {
-                e.get_mut().extend(items);
-            }
+            i.defer_processing
         });
-        self.set_timer_if_required();
+
+        if defer_processing {
+            self.set_timer_if_required();
+        } else {
+            self.run();
+        }
     }
 
     fn set_timer_if_required(&self) -> bool {
@@ -111,18 +133,25 @@ where
                         }
 
                         let mut batch = T::new(grouping_key);
+                        let mut empty_batch = true;
                         let queue = e.get_mut();
-                        while !batch.is_full() {
+                        loop {
                             if let Some(next) = queue.pop_front() {
                                 batch.add(next);
+                                empty_batch = false;
                             } else {
+                                break;
+                            }
+                            if batch.is_full() {
                                 break;
                             }
                         }
                         if queue.is_empty() {
                             e.remove();
                         }
-                        batches.push(batch);
+                        if !empty_batch {
+                            batches.push(batch);
+                        }
                     }
                 } else {
                     break;